SceneDelegate has been deleted...
Now it starts from AppDelegate!
Set up folders in project for convenience.
