+++ YandexWeatherTestWork/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
