+++ YandexWeatherTestWork/Resources/Assets.xcassets/Contents.json
